feat(submitentityclaim): finish singledateselector
