+++ src/common/components/JsonForm/CustomWidgets/SingleDateSelector/SingleDateSelector.tsx
@@ -1,0 +1,43 @@
+import React from 'react'
+import { SingleDatePicker } from 'react-dates'
+import moment, { Moment } from 'moment'
+
+interface Props {
+  id: string
+  value: string
+  onChange: (value: string) => void
+}
+
+interface State {
+  focused: boolean
+}
+
+class SingleDateSelector extends React.Component<Props, State> {
+  constructor(props) {
+    super(props)
+
+    this.state = {
+      focused: false,
+    }
+  }
+
+  render(): JSX.Element {
+    const { id, value, onChange } = this.props
+
+    return (
+      <SingleDatePicker
+        date={value ? moment(value) : null}
+        displayFormat="DD-MMM-YYYY"
+        onDateChange={(date: Moment): void =>
+          onChange(date ? date.format('DD-MMM-YYYY') : null)
+        }
+        focused={this.state.focused}
+        onFocusChange={({ focused }): void => this.setState({ focused })}
+        id={id}
+        isOutsideRange={(): boolean => false}
+      />
+    )
+  }
+}
+
+export default SingleDateSelector
