+++ src/common/components/JsonForm/SingleControlForm/SingleControlForm.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Form from '@rjsf/core'
 import { FormContainer } from './SingleControlForm.styles'
 import ImageCheckboxes from '../CustomWidgets/ImageCheckboxes/ImageCheckboxes'
-import SingleDatePicker from '../CustomWidgets/SingleDatePicker/SingleDatePicker'
+import SingleDateSelector from '../CustomWidgets/SingleDateSelector/SingleDateSelector'
 import { FormControl } from '../types'
 
 interface Props {
@@ -15,7 +15,7 @@ interface Props {
 
 const customControls = {
   ['imagecheckboxes']: ImageCheckboxes,
-  ['singledatepicker']: SingleDatePicker,
+  ['singledateselector']: SingleDateSelector,
 }
 
 const SingleControlForm: React.FunctionComponent<Props> = ({
